refactor: 기존의 컨트롤러의 playRound함수를 Participants클래스에게 위임
